requirements.txt
python-telegram-bot==13.13
openai==1.12.0
supabase==2.3.0
python-dotenv==1.0.0
pytz==2023.3
pillow==10.0.0
requests==2.31.0
datetime==5.2
uuid==1.30